# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Fleet Operations API
  version: 1.0.0
  description: 车队经营管理系统统一 API 契约。详细业务规则以项目需求基线文档为准。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      tags: [System]
      summary: 用户分页列表
      x-permission: user:manage
      responses: {'200': {description: OK}}
    post:
      tags: [System]
      summary: 新增账号并分配单一角色
      x-permission: user:manage
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UserRequest'}}}}
      responses: {'200': {description: OK}, '409': {$ref: '#/components/responses/BusinessConflict'}}
  /users/{id}:
    get:
      tags: [System]
      summary: 用户详情
      x-permission: user:manage
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
    put:
      tags: [System]
      summary: 编辑账号并保持单一角色
      x-permission: user:manage
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UserRequest'}}}}
      responses: {'200': {description: OK}, '409': {$ref: '#/components/responses/BusinessConflict'}}
  /roles:
    get:
      tags: [System]
      summary: 角色分页列表
      x-permission: role:manage
      responses: {'200': {description: OK}}
    post:
      tags: [System]
      summary: 新增角色
      x-permission: role:manage
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RoleRequest'}}}}
      responses: {'200': {description: OK}, '409': {$ref: '#/components/responses/BusinessConflict'}}
  /roles/{id}:
    get:
      tags: [System]
      summary: 角色详情
      x-permission: role:manage
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
    put:
      tags: [System]
      summary: 编辑角色
      x-permission: role:manage
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RoleRequest'}}}}
      responses: {'200': {description: OK}, '409': {$ref: '#/components/responses/BusinessConflict'}}
  /roles/{id}/status:
    put:
      tags: [System]
      summary: 启用/停用角色
      x-permission: role:manage
      parameters:
        - {$ref: '#/components/parameters/Id'}
        - {name: status, in: query, required: true, schema: {type: integer, enum: [0, 1]}}
      responses: {'200': {description: OK}}
  /roles/{id}/permissions:
    put:
      tags: [System]
      summary: 保存角色权限；权限码只能来自既定注册表
      x-permission: role:manage
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RolePermissionRequest'}}}}
      responses: {'200': {description: OK}, '409': {$ref: '#/components/responses/BusinessConflict'}}
  /permissions:
    get:
      tags: [System]
      summary: 权限列表，支持按 MENU/BUTTON/API/MINI_PROGRAM 分类
      x-permission: permission:manage
      parameters:
        - {name: permissionType, in: query, required: false, schema: {type: string, enum: [MENU, BUTTON, API, MINI_PROGRAM]}}
      responses: {'200': {description: OK}}
  /audit-logs:
    get:
      tags: [System]
      summary: 操作日志分页查询
      x-permission: audit:view
      parameters:
        - {name: module, in: query, required: false, schema: {type: string}}
        - {name: businessType, in: query, required: false, schema: {type: string}}
        - {name: operatorId, in: query, required: false, schema: {type: integer, format: int64}}
        - {name: startTime, in: query, required: false, schema: {type: string, format: date-time}}
        - {name: endTime, in: query, required: false, schema: {type: string, format: date-time}}
      responses: {'200': {description: OK}}
...
  /tire-requests/{id}:
    get:
      tags: [Tires]
      summary: 换胎申请详情
      x-permission: tire:approval:view
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses: {'200': {description: OK}}
  /ocr/tire-number:
    post:
      tags: [OCR]
      summary: 识别胎号
      description: 后端根据 attachmentId 读取 TIRE_OCR 原图，Base64 后调用腾讯云 GeneralAccurateOCR；不接受调用方传 recognizedText/rawResultJson 作为识别结果来源。
      x-permission: tire:request
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OcrTireNumberRequest'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - {$ref: '#/components/schemas/ApiResponse'}
                  - type: object
                    properties:
                      data: {$ref: '#/components/schemas/OcrRecord'}
  /ocr/{id}/confirm:
    post:
      tags: [OCR]
      summary: 确认 OCR 最终胎号
      x-permission: tire:request
      parameters: [{$ref: '#/components/parameters/Id'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OcrConfirmRequest'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - {$ref: '#/components/schemas/ApiResponse'}
                  - type: object
                    properties:
                      data: {$ref: '#/components/schemas/OcrRecord'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: integer, format: int64}}
...
  responses:
    BusinessConflict:
      description: 业务规则冲突
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiResponse'}
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code: {type: string, examples: ['0']}
        message: {type: string}
        data: {}
        requestId: {type: string}
...
    UserRequest:
      type: object
      required: [phone, roleId, status]
      properties:
        phone: {type: string}
        password: {type: [string, 'null'], description: 新增账号必填；编辑账号不使用该字段重置密码}
        roleId: {type: integer, format: int64}
        status: {type: integer, enum: [0, 1]}
...
    RoleRequest:
      type: object
      required: [roleCode, roleName, status]
      properties:
        roleCode: {type: string}
        roleName: {type: string}
        status: {type: integer, enum: [0, 1]}
    RolePermissionRequest:
      type: object
      properties:
        permissionCodes:
          type: array
          items: {$ref: '#/components/schemas/PermissionCode'}
...
    OcrTireNumberRequest:
      type: object
      required: [attachmentId]
      properties:
        attachmentId:
          type: integer
          format: int64
          description: 已上传的 TIRE_OCR 原图附件 ID。后端读取附件 bytes 并调用 OCR Provider。
    OcrConfirmRequest:
      type: object
      required: [confirmedText]
      properties:
        confirmedText: {type: string, minLength: 1, description: 司机最终确认的胎号，可来自候选或人工输入。}
    OcrCandidate:
      type: object
      properties:
        candidate: {type: string}
        confidence: {type: [number, 'null'], format: double}
        sourceText: {type: string}
        inventoryMatched: {type: boolean}
    OcrRecord:
      type: object
      properties:
        id: {type: integer, format: int64}
        attachmentId: {type: integer, format: int64}
        ocrProvider: {type: [string, 'null'], example: TENCENT_GENERAL_ACCURATE}
        providerRequestId: {type: [string, 'null']}
        rawResultJson: {type: [string, 'null'], description: 腾讯云 GeneralAccurateOCR 原始响应 JSON，不包含密钥。}
        recognizedText: {type: [string, 'null'], description: OCR 检测到的原始文本聚合。}
        candidateText: {type: [string, 'null'], description: 排序最高的胎号候选。}
        ocrStatus: {type: string, enum: [SUCCESS, NO_CANDIDATE, FAILED]}
        errorCode: {type: [string, 'null']}
        errorMessage: {type: [string, 'null']}
        candidates:
          type: array
          items: {$ref: '#/components/schemas/OcrCandidate'}
        confirmedText: {type: [string, 'null']}
        confirmedBy: {type: [integer, 'null'], format: int64}
        confirmedAt: {type: [string, 'null'], format: date-time}
...
    PermissionCode:
      type: string
      enum:
        - vehicle:list
        - vehicle:add
        - vehicle:view
        - vehicle:edit
        - vehicle:delete
        - vehicle:bindDriver
        - vehicle:unbindDriver
        - vehicle:bindingHistory
        - trailer:list
        - trailer:add
        - trailer:edit
        - trailer:bindVehicle
        - trailer:unbindVehicle
        - trailer:bindingHistory
        - driver:list
        - driver:add
        - driver:edit
        - driver:bindVehicle
        - driver:unbindVehicle
        - driver:resetPassword
        - route:list
        - route:create
        - route:edit
        - route:publish
        - route:cancel
        - route:delete
        - route:depart
        - route:unload
        - route:void
        - route:reactivate
        - route:weight:adjust
        - expense:list
        - expense:add
        - expense:edit
        - expense:delete
        - expense:attribution:edit
        - expense:reversal
        - expense:approval:view
        - approval:create
        - approval:process
        - approval:return
        - approval:history:view
        - approval:flow:manage
        - tire:list
        - tire:import
        - tire:request
        - tire:approval:view
        - import:preview
        - import:commit
        - import:history
        - import:failure:export
        - salary:view
        - salary:manage
        - salary:mine
        - report:dashboard
        - report:profit
        - report:vehicle
        - report:driver
        - report:attendance
        - report:energy
        - settlement:view
        - settlement:generate
        - settlement:diff:view
        - user:manage
        - role:manage
        - permission:manage
        - audit:view